rename(route-folder): 1:1 질문 작성 컨벤션 수정
