--- src/app/(route)/inquiry-write/_hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useInquiryWrite } from "./useInquiryWrite/useInquiryWrite";
